\name{iGen-package}
\alias{iGen-package}
\alias{iGen}
\docType{package}
\title{
\packageTitle{iGen}
}
\description{
\packageDescription{iGen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iGen}
\packageIndices{iGen}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{iGen}

Maintainer: \packageMaintainer{iGen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
